+++ src/controller/customerController.ts
@@ -9,101 +9,6 @@ import { UserSignInDTO } from "../interfaces/user/userSignInDTO";
 import { CustomerUpdateDTO } from "../interfaces/user/customerUpdateDTO";
 import { CreateDeliveryRequestDTO } from "../interfaces/delivery/createDeliveryRequestDTO";
 
-// 고객 내 근처 카페 개별 업체 정보 조회
-const getNearCafeById = async (req: Request, res: Response) => {
-  const storeId = req.params.id;
-  if (!storeId) {
-    return res.status(sc.BAD_REQUEST).send(fail(sc.BAD_REQUEST, rm.NULL_VALUE));
-  }
-  try {
-    const data = await customerService.getNearCafeById(+storeId);
-    if (!data) {
-      return res
-        .status(sc.BAD_REQUEST)
-        .send(fail(sc.BAD_REQUEST, rm.GET_NEAR_CAFE_FAIL));
-    }
-    return res
-      .status(sc.OK)
-      .send(success(sc.OK, rm.GET_NEAR_CAFE_SUCCESS, data));
-  } catch (error) {
-    console.log(error);
-    res
-      .status(sc.INTERNAL_SERVER_ERROR)
-      .send(fail(sc.INTERNAL_SERVER_ERROR, rm.INTERNAL_SERVER_ERROR));
-  }
-};
-
-// 고객 근처 카페 개별 업체 소식 조회
-const getCafeNoticeById = async (req: Request, res: Response) => {
-  const id = req.params.id;
-  if (!id) {
-    return res.status(sc.BAD_REQUEST).send(fail(sc.BAD_REQUEST, rm.NULL_VALUE));
-  }
-  try {
-    const data = await customerService.getCafeNoticeById(+id);
-    if (!data || data.length === 0) {
-      return res
-        .status(sc.BAD_REQUEST)
-        .send(fail(sc.BAD_REQUEST, rm.GET_CAFE_NOTICE_FAIL));
-    }
-    return res
-      .status(sc.OK)
-      .send(success(sc.OK, rm.GET_CAFE_NOTICE_SUCCESS, data));
-  } catch (error) {
-    console.log(error);
-    res
-      .status(sc.INTERNAL_SERVER_ERROR)
-      .send(fail(sc.INTERNAL_SERVER_ERROR, rm.INTERNAL_SERVER_ERROR));
-  }
-};
-
-// 고객 근처 카페 개별 업체 메뉴 조회
-const getCafeMenuById = async (req: Request, res: Response) => {
-  const id = req.params.id;
-  if (!id) {
-    return res.status(sc.BAD_REQUEST).send(fail(sc.BAD_REQUEST, rm.NULL_VALUE));
-  }
-  try {
-    const data = await customerService.getCafeMenuById(+id);
-    if (!data) {
-      return res
-        .status(sc.BAD_REQUEST)
-        .send(fail(sc.BAD_REQUEST, rm.GET_CAFE_MENU_FAIL));
-    }
-    return res
-      .status(sc.OK)
-      .send(success(sc.OK, rm.GET_CAFE_MENU_SUCCESS, data));
-  } catch (error) {
-    console.log(error);
-    res
-      .status(sc.INTERNAL_SERVER_ERROR)
-      .send(fail(sc.INTERNAL_SERVER_ERROR, rm.INTERNAL_SERVER_ERROR));
-  }
-};
-
-// 고객 근처 카페 개별 업체 피드 조회
-const getCafeReviewById = async (req: Request, res: Response) => {
-  const storeId = req.params.id;
-  if (!storeId) {
-    return res.status(sc.BAD_REQUEST).send(fail(sc.BAD_REQUEST, rm.NULL_VALUE));
-  }
-  try {
-    const data = await customerService.getCafeReviewById(+storeId);
-    if (!data || data.length === 0) {
-      return res
-        .status(sc.BAD_REQUEST)
-        .send(fail(sc.BAD_REQUEST, rm.GET_CAFE_REVIEW_FAIL));
-    }
-    return res
-      .status(sc.OK)
-      .send(success(sc.OK, rm.GET_CAFE_REVIEW_SUCCESS, data));
-  } catch (error) {
-    console.log(error);
-    res
-      .status(sc.INTERNAL_SERVER_ERROR)
-      .send(fail(sc.INTERNAL_SERVER_ERROR, rm.INTERNAL_SERVER_ERROR));
-  }
-};
 // 고객 스탬프 사용 신청
 const createDeliveryRequest = async (req: Request, res: Response) => {
   // validation의 결과를 바탕으로 분기 처리
@@ -400,10 +305,6 @@ const customerController = {
   updateCustomer,
   customerDelete,
   getCustomerName,
-  getNearCafeById,
-  getCafeNoticeById,
-  getCafeMenuById,
-  getCafeReviewById,
   createStampNumber,
   createDeliveryRequest,
   getAllStamp,
