+++ src/controller/index.ts
@@ -1,3 +1,4 @@
 export { default as customerController } from "./customerController";
 export { default as ownerController } from "./ownerController";
 export { default as managerController } from "./managerController";
+export { default as mainController } from "./mainController";
