+++ src/router/customerRouter.ts
@@ -5,18 +5,6 @@ import { auth } from "../middlewares";
 
 const router: Router = Router();
 
-// 고객 근처 카페 개별 업체 정보 조회
-router.get("/store/info/:id", auth, customerController.getNearCafeById);
-
-// 고객 근처 카페 개별 업체 소식 조회
-router.get("/store/notice/:id", auth, customerController.getCafeNoticeById);
-
-// 고객 근처 카페 개별 업체 메뉴 조회
-router.get("/store/menu/:id", auth, customerController.getCafeMenuById);
-
-// 고객 근처 카페 개별 업체 피드 조회
-router.get("/store/review/:id", auth, customerController.getCafeReviewById);
-
 // 고객 스탬프 사용신청 = POST ~/customer/delivery
 router.post(
   "/delivery",
