+++ src/service/index.ts
@@ -1,3 +1,4 @@
 export { default as customerService } from "./customerService";
 export { default as ownerService } from "./ownerService";
 export { default as managerService } from "./managerService";
+export { default as mainService } from "./mainService";
